requirements1.txt
amqp==5.3.1
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.8.2
celery==5.5.3
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
cron_descriptor==2.0.6
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.0.14
django-appconf==1.1.0
django-celery-beat==2.8.1
django-crispy-forms==2.4
django-redis==6.0.0
django-select2==8.4.1
django-timezone-field==7.1
django-widget-tweaks==1.5.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.58.4
html5lib==1.1
idna==3.11
importlib_resources==6.5.2
kombu==5.5.4
mssql-django==1.5
mysqlclient==2.2.7
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.2.1
prompt_toolkit==3.0.52
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
PyMySQL==1.1.2
pyodbc==5.2.0
pyotp==2.9.0
PyPDF2==3.0.1
pyphen==0.17.2
pypng==0.20220715.0
python-barcode==0.15.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-escpos==3.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
qrcode==7.4.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.5
schedule==1.2.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==6.11.0
win32printing==0.1.3
zopfli==0.2.3.post1
